requirements.txt
llama-cpp-python
python-dotenv
pydantic